[feat]: change max log file size type from int to long

+++ firefly-common/src/main/java/com/firefly/utils/log/XmlLogConfigParser.java
@@ -30,10 +30,8 @@ public class XmlLogConfigParser extends AbstractLogConfigParser {
 				String name = dom.getTextValueByTagName(e, "name", DEFAULT_LOG_NAME);
 				String level = dom.getTextValueByTagName(e, "level", DEFAULT_LOG_LEVEL);
 				String path = dom.getTextValueByTagName(e, "path", DEFAULT_LOG_DIRECTORY.getAbsolutePath());
-				boolean consoleEnabled = ConvertUtils.convert(dom.getTextValueByTagName(e, "enable-console"),
-						DEFAULT_CONSOLE_ENABLED);
-				int maxFileSize = ConvertUtils.convert(dom.getTextValueByTagName(e, "max-file-size"),
-						DEFAULT_MAX_FILE_SIZE);
+				boolean consoleEnabled = ConvertUtils.convert(dom.getTextValueByTagName(e, "enable-console"), DEFAULT_CONSOLE_ENABLED);
+				long maxFileSize = ConvertUtils.convert(dom.getTextValueByTagName(e, "max-file-size"), DEFAULT_MAX_FILE_SIZE);
 				if (maxFileSize < 1024 * 1024 * 10) {
 					System.err.println("the max log file less than 10MB, please set a larger file size");
 				}
